+++ Cookie/Game/Entity/Entity.cs
@@ -4,13 +4,13 @@ namespace Cookie.Game.Entity
 {
     public class Entity : IEntity
     {
-        public Entity(int id, int cellId)
+        public Entity(double id, int cellId)
         {
             Id = id;
             CellId = cellId;
         }
 
         public int CellId { get; set; }
-        public int Id { get; set; }
+        public double Id { get; set; }
     }
 }
